{"schema":"decodex/commit/1","summary":"Modularize recovery context and runtime sources","authority":"manual"}
